refactor: split dynamic_writer from writer
